Refactor api.py: extract Trigger + public webhook routes (#367)
Continues the api.py split. Stacks on top of #366.

- /triggers + /agents/{name}/triggers/* (8 endpoints) → routes/triggers.py
- /hooks/{token} public webhook receiver → routes/triggers.py

The IP rate-limit and per-token rate-limit buckets, plus the
{{body.field}} prompt-template renderer, all moved into the new module
along with the routes that use them. The wake helper stayed in api.py
(it has deep dependencies on broker / streaming sessions / per-agent
state) and is passed into routes/triggers.py via set_dependencies as
a callable.

api.py: 8,389 → 8,153 lines (-236, -2.8%)
Cumulative since main: 11,138 → 8,153 (-2,985 lines, -26.8%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
